split client code into AMD modules and load them with requirejs
