+++ js/OC.js
@@ -1,0 +1,18 @@
+/* global OC */
+
+/**
+ * ownCloud - Mail
+ *
+ * This file is licensed under the Affero General Public License version 3 or
+ * later. See the COPYING file.
+ *
+ * @author Christoph Wurst <christoph@winzerhof-wurst.at>
+ * @copyright Christoph Wurst 2015
+ */
+
+define(function() {
+	'use strict';
+
+	return OC;
+});
+
