+++ js/models/messageflags.js
@@ -11,17 +11,11 @@
 define(function(require) {
 	'use strict';
 
-	var Marionette = require('marionette');
+	var Backbone = require('backbone');
 
-	return Marionette.ItemView.extend({
-		tagName: 'li',
-		template: '#mail-attachment-template',
-		events: {
-			'click .icon-delete': 'removeAttachment'
-		},
-		removeAttachment: function() {
-			this.model.collection.remove(this.model);
+	return Backbone.Model.extend({
+		defaults: {
+			answered: false
 		}
-
 	});
 });
