+++ js/views/helper.js
@@ -1,5 +1,3 @@
-/* global Mail, Handlebars, _, relative_modified_date, formatDate, md5, humanFileSize, getScrollBarWidth */
-
 /**
  * ownCloud - Mail
  *
@@ -10,7 +8,11 @@
  * @copyright Christoph Wurst 2015
  */
 
-$(document).ready(function() {
+define(function(require) {
+	'use strict';
+
+	var Handlebars = require('handlebars');
+
 	Handlebars.registerHelper("relativeModifiedDate", function(dateInt) {
 		var lastModified = new Date(dateInt * 1000);
 		var lastModifiedTime = Math.round(lastModified.getTime() / 1000);
@@ -26,16 +28,16 @@ $(document).ready(function() {
 		return humanFileSize(size);
 	});
 
-	Handlebars.registerHelper('accountColor', function (account) {
+	Handlebars.registerHelper('accountColor', function(account) {
 		var hash = md5(account),
 			maxRange = parseInt('ffffffffffffffffffffffffffffffff', 16),
 			hue = parseInt(hash, 16) / maxRange * 256;
 		return new Handlebars.SafeString('hsl(' + hue + ', 90%, 65%)');
 	});
 
-	Handlebars.registerHelper("printAddressList", function (addressList) {
-		var currentAddress = _.find(Mail.State.accounts, function (item) {
-			return item.accountId === Mail.State.currentAccountId;
+	Handlebars.registerHelper("printAddressList", function(addressList) {
+		var currentAddress = _.find(require('app').State.accounts, function(item) {
+			return item.accountId === require('app').State.currentAccountId;
 		});
 
 		var str = _.reduce(addressList, function(memo, value, index) {
@@ -52,7 +54,7 @@ $(document).ready(function() {
 			}
 			var title = t('mail', 'Send message to {email}', {email: email});
 			memo += '<span class="tipsy-mailto" title="' + title + '">';
-			memo += '<a class="link-mailto" data-email="' +	email + '" data-label="' +	label + '">';
+			memo += '<a class="link-mailto" data-email="' + email + '" data-label="' + label + '">';
 			memo += label + '</a></span>';
 			return memo;
 		}, "");
@@ -78,7 +80,7 @@ $(document).ready(function() {
 		return str;
 	});
 
-	Handlebars.registerHelper("ifHasCC", function (cc, ccList, options) {
+	Handlebars.registerHelper("ifHasCC", function(cc, ccList, options) {
 		if (!_.isUndefined(cc) || (!_.isUndefined(ccList) && ccList.length > 0)) {
 			return options.fn(this);
 		} else {
@@ -86,7 +88,7 @@ $(document).ready(function() {
 		}
 	});
 
-	Handlebars.registerHelper("unlessHasCC", function (cc, ccList, options) {
+	Handlebars.registerHelper("unlessHasCC", function(cc, ccList, options) {
 		if (_.isUndefined(cc) && (_.isUndefined(ccList) || ccList.length === 0)) {
 			return options.fn(this);
 		} else {
@@ -109,27 +111,28 @@ $(document).ready(function() {
 		outer.style.width = "200px";
 		outer.style.height = "150px";
 		outer.style.overflow = "hidden";
-		outer.appendChild (inner);
+		outer.appendChild(inner);
 
-		document.body.appendChild (outer);
+		document.body.appendChild(outer);
 		var w1 = inner.offsetWidth;
 		outer.style.overflow = 'scroll';
 		var w2 = inner.offsetWidth;
-		if(w1 === w2) {
+		if (w1 === w2) {
 			w2 = outer.clientWidth;
 		}
 
-		document.body.removeChild (outer);
+		document.body.removeChild(outer);
 
 		return (w1 - w2);
 	};
 	//END TODO
 
+	// TODO: get rid of global functions
 	// adjust controls/header bar width
 	window.adjustControlsWidth = function() {
-		if($('#mail-message-header').length) {
+		if ($('#mail-message-header').length) {
 			var controlsWidth;
-			if($(window).width() > 768) {
+			if ($(window).width() > 768) {
 				controlsWidth =
 					$('#content').width() -
 					$('#app-navigation').width() -
@@ -142,5 +145,6 @@ $(document).ready(function() {
 	};
 
 	$(window).resize(_.debounce(window.adjustControlsWidth, 250));
-
+	// END TODO
 });
+
